# [Enhancement] Optimize a subtle inline performance problem